MVVM

+++ app/src/main/java/com/cs4520/assignment3/Presenter.kt
@@ -5,9 +5,9 @@ import java.lang.NumberFormatException
 class Presenter (private var View: MVPContract.View?) : MVPContract.Presenter{
     override fun OnAddClick(num1: String, num2: String) {
         try {
-            val n1 = num1.toInt()
-            val n2 = num2.toInt()
-            View?.ShowResult((n1 + n2).toFloat())
+            val n1 = num1.toFloat()
+            val n2 = num2.toFloat()
+            View?.ShowResult((n1 + n2))
         } catch (e : NumberFormatException){
             View?.ShowToast()
         }
@@ -15,9 +15,9 @@ class Presenter (private var View: MVPContract.View?) : MVPContract.Presenter{
 
     override fun OnSubtractClick(num1: String, num2: String) {
         try {
-            val n1 = num1.toInt()
-            val n2 = num2.toInt()
-            View?.ShowResult((n1 - n2).toFloat())
+            val n1 = num1.toFloat()
+            val n2 = num2.toFloat()
+            View?.ShowResult((n1 - n2))
         } catch (e : NumberFormatException){
             View?.ShowToast()
         }
@@ -26,9 +26,9 @@ class Presenter (private var View: MVPContract.View?) : MVPContract.Presenter{
     override fun OnMultiplyClick(num1: String, num2: String) {
 
         try {
-            val n1 = num1.toInt()
-            val n2 = num2.toInt()
-            View?.ShowResult((n1 * n2).toFloat())
+            val n1 = num1.toFloat()
+            val n2 = num2.toFloat()
+            View?.ShowResult((n1 * n2))
         } catch (e : NumberFormatException){
             View?.ShowToast()
         }
